[httpjson] Add value_type parameter to httpjson transforms (#26847)
* [httpjson] Add new tpl functions

* [httpjson] Change last_response.page internal type to int64

* [httpjson] Add new value_type parameter

* Add changelog entry

* Add new value type to tests

+++ x-pack/filebeat/input/httpjson/internal/v2/response.go
@@ -22,7 +22,7 @@ func registerResponseTransforms() {
 }
 
 type response struct {
-	page   int
+	page   int64
 	url    url.URL
 	header http.Header
 	body   interface{}
